Fixed attack cooldown bug
Bug details: player could only attack once and cooldown timer wasn't working properly as was only decreasing when the correct keybind was down. Fixed the issue by moving the cooldown time decrease code out of the keybind code (no idea why it was there in the first place) this means it can now run in the update method independent of the attack code.