Fixed player sprite positioning issues.
The player was appearing to be underneath objects so it was changed so the player was slightly further off the screen than the rest so it would be above them.

This has created a bug with the player being above monsters and other objects making it look as though the player is on top of them.