Slight fix for object positioning bug
Changed positioning so that objects like the traps were back slightly, and the player was back on the same level. This means that the player will still be above the traps but not standing on other objects like enemies.

However this means that when the player is positioned 'below' the enemies they now appear to stand on top of the player. This issue will need to be fixed with some code that prioritizes the z positioning of some objects based on their y level compared to the player.