Fixed canvas ordering bug
When UI elements were spawned in like health bars for enemies they would be on top of other UI elements like the inventory. This was fixed by adding an empty object ordered at the top of the canvas so it would be below everything that object could be spawned into.